+++ src/worker.rs
@@ -211,7 +211,6 @@ pub async fn worker_main(goose_attack: &GooseAttack) -> GooseAttack {
     }
 
     // Worker is officially starting the load test.
-    let started = time::Instant::now();
     info!(
         "[{}] entering gaggle mode, starting load test",
         get_worker_id()
@@ -219,6 +218,7 @@ pub async fn worker_main(goose_attack: &GooseAttack) -> GooseAttack {
     let sleep_duration = time::Duration::from_secs_f32(hatch_rate.unwrap());
 
     let mut worker_goose_attack = GooseAttack::initialize_with_config(config.clone());
+    worker_goose_attack.started = Some(time::Instant::now());
     worker_goose_attack.task_sets = goose_attack.task_sets.clone();
     if config.run_time != "" {
         worker_goose_attack.run_time = util::parse_timespan(&config.run_time);
@@ -233,7 +233,7 @@ pub async fn worker_main(goose_attack: &GooseAttack) -> GooseAttack {
     worker_goose_attack.weighted_users = weighted_users;
     worker_goose_attack.configuration.worker = true;
     worker_goose_attack
-        .launch_users(started, sleep_duration, Some(manager))
+        .launch_users(sleep_duration, Some(manager))
         .await
 }