+++ tests/logs.rs
@@ -48,7 +48,7 @@ fn test_stat_logs_json() {
     let mut config = common::build_configuration();
     config.stats_log_file = STATS_LOG_FILE.to_string();
     config.no_stats = false;
-    crate::GooseAttack::initialize_with_config(config)
+    let _stats = crate::GooseAttack::initialize_with_config(config)
         .setup()
         .register_taskset(taskset!("LoadTest").register_task(task!(get_index)))
         .execute();
@@ -76,7 +76,7 @@ fn test_stat_logs_csv() {
     config.stats_log_file = STATS_LOG_FILE.to_string();
     config.stats_log_format = "csv".to_string();
     config.no_stats = false;
-    crate::GooseAttack::initialize_with_config(config)
+    let _stats = crate::GooseAttack::initialize_with_config(config)
         .setup()
         .register_taskset(taskset!("LoadTest").register_task(task!(get_index)))
         .execute();
@@ -104,7 +104,7 @@ fn test_stat_logs_raw() {
     config.stats_log_file = STATS_LOG_FILE.to_string();
     config.stats_log_format = "raw".to_string();
     config.no_stats = false;
-    crate::GooseAttack::initialize_with_config(config)
+    let _stats = crate::GooseAttack::initialize_with_config(config)
         .setup()
         .register_taskset(taskset!("LoadTest").register_task(task!(get_index)))
         .execute();
@@ -132,7 +132,7 @@ fn test_debug_logs_raw() {
     let mut config = common::build_configuration();
     config.debug_log_file = DEBUG_LOG_FILE.to_string();
     config.debug_log_format = "raw".to_string();
-    crate::GooseAttack::initialize_with_config(config)
+    let _stats = crate::GooseAttack::initialize_with_config(config)
         .setup()
         .register_taskset(
             taskset!("LoadTest")
@@ -165,7 +165,7 @@ fn test_debug_logs_json() {
 
     let mut config = common::build_configuration();
     config.debug_log_file = DEBUG_LOG_FILE.to_string();
-    crate::GooseAttack::initialize_with_config(config)
+    let _stats = crate::GooseAttack::initialize_with_config(config)
         .setup()
         .register_taskset(
             taskset!("LoadTest")
@@ -201,7 +201,7 @@ fn test_stats_and_debug_logs() {
     config.stats_log_format = "raw".to_string();
     config.no_stats = false;
     config.debug_log_file = DEBUG_LOG_FILE.to_string();
-    crate::GooseAttack::initialize_with_config(config)
+    let _stats = crate::GooseAttack::initialize_with_config(config)
         .setup()
         .register_taskset(
             taskset!("LoadTest")